requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver
redis==5.0.1
asyncio-redis==0.16.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP requests and async
httpx==0.25.2
aiohttp==3.9.1

# Rate limiting and caching
slowapi==0.1.9
cachetools==5.3.2

# Data processing and ML
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
joblib==1.3.2

# Time series analysis for pattern detection
statsmodels==0.14.0
plotly==5.17.0

# Message queuing (optional)
celery==5.3.4
kafka-python==2.0.2

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # for testing FastAPI
faker==20.1.0  # for generating test data

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Environment management
python-dotenv==1.0.0

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# JSON handling
orjson==3.9.10

# Validation
email-validator==2.1.0
phonenumbers==8.13.25

# Performance monitoring
psutil==5.9.6

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL driver
alembic==1.13.1         # Database migrations
redis==5.0.1            # Redis client
